+++ packages/vendure-plugin-utm-tracker/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added UI to view UTM parameters on the order detail page.
 - Added U-Shaped Attribution model.
+- Reversed orders of UTM parameters in API: Most recent parameters are now first.
 
 # 1.1.1 (2025-10-27)